+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
-- Add drop-down filters for constellation and rarity to extractions (#12)
+- Add drop-down filters for constellation to extractions (#12)
+- Add drop-down filters for rarity to extractions (#12)
+- Add drop-down filters for constellation to moons (#12)
 
 ### Changed